requirements.txt
streamlit
tensorflow
opencv-python-headless
scikit-learn
numpy
Pillow
pandas
matplotlib
seaborn